Move common test configurations to base test classes to avoid running tests twice
