+++ .travis.yml
@@ -47,7 +47,6 @@ script:
   - chmod 755 gradlew
   - ls -la
   #Run jacocoTestReport -> Runs all unit tests & instrumentation tests
-  - ./gradlew :rdt-open-reader:clean :rdt-open-reader:testMalariaIndonesiaDebugUnitTestCoverage
   - ./gradlew :rdt-open-reader:clean :rdt-open-reader:testMalariaKenyaDebugUnitTestCoverage
   - ./gradlew :rdt-open-reader:clean :rdt-open-reader:testCovidIndonesiaDebugUnitTestCoverage
   - ./gradlew coveralls --stacktrace --info
