+++ rdt-open-reader/src/test/java/io/ona/rdt/presenter/RDTJsonFormFragmentPresenterTest.java
@@ -1,47 +1,20 @@
 package io.ona.rdt.presenter;
 
-import android.content.Context;
-
 import com.vijay.jsonwizard.fragments.JsonFormFragment;
 import com.vijay.jsonwizard.views.JsonFormFragmentView;
 import com.vijay.jsonwizard.viewstates.JsonFormFragmentViewState;
 
-import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.AdditionalMatchers;
 import org.mockito.Mock;
-import org.powermock.api.mockito.PowerMockito;
-import org.powermock.core.classloader.annotations.PrepareForTest;
-import org.powermock.modules.junit4.PowerMockRunner;
 import org.powermock.reflect.Whitebox;
-import org.smartregister.repository.AllSharedPreferences;
-import org.smartregister.repository.EventClientRepository;
 
 import java.lang.ref.WeakReference;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
-import io.ona.rdt.application.RDTApplication;
-import io.ona.rdt.contract.RDTJsonFormFragmentContract;
-import io.ona.rdt.fragment.RDTJsonFormFragment;
-import io.ona.rdt.interactor.RDTJsonFormFragmentInteractor;
-import io.ona.rdt.interactor.RDTJsonFormInteractor;
-import io.ona.rdt.util.StepStateConfig;
-
 import static io.ona.rdt.TestUtils.getDateWithOffset;
-import static io.ona.rdt.util.Constants.Step.BLOT_PAPER_TASK_PAGE;
-import static io.ona.rdt.util.Constants.Step.EXPIRATION_DATE_READER_ADDRESS;
-import static io.ona.rdt.util.Constants.Step.MANUAL_EXPIRATION_DATE_ENTRY_PAGE;
-import static io.ona.rdt.util.Constants.Step.PRODUCT_EXPIRED_PAGE;
-import static io.ona.rdt.util.Constants.Step.RDT_ID_KEY;
-import static io.ona.rdt.util.Constants.Step.RDT_ID_LBL_ADDRESSES;
-import static io.ona.rdt.util.Constants.Step.SCAN_CARESTART_PAGE;
-import static io.ona.rdt.util.Constants.Step.SCAN_QR_PAGE;
-import static io.ona.rdt.util.Constants.Step.TAKE_IMAGE_OF_RDT_PAGE;
 import static io.ona.rdt.util.RDTJsonFormUtilsTest.RDT_TEST_JSON_FORM;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
@@ -51,45 +24,16 @@ import static org.mockito.ArgumentMatchers.isNull;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
-import static org.powermock.api.mockito.PowerMockito.mockStatic;
-import static org.powermock.api.mockito.PowerMockito.when;
 
 /**
  * Created by Vincent Karuri on 14/08/2019
  */
-@RunWith(PowerMockRunner.class)
-@PrepareForTest({RDTJsonFormFragment.class, RDTApplication.class})
-public class RDTJsonFormFragmentPresenterTest {
 
-    protected RDTJsonFormFragmentPresenter presenter;
-    protected RDTJsonFormFragmentContract.View rdtFormFragmentView;
-    protected RDTJsonFormFragment formFragment;
+public class RDTJsonFormFragmentPresenterTest extends BaseRDTJsonFormFragmentPresenterTest {
+
+    @Mock
     private JsonFormFragmentView<JsonFormFragmentViewState> view;
-
-    @Mock
-    private RDTJsonFormFragmentInteractor interactor;
-
-    @Mock
-    private RDTApplication rdtApplication;
-
-    @Mock
-    private StepStateConfig stepStateConfig;
-
-    @Mock
-    private org.smartregister.Context drishtiContext;
-
-    @Mock
-    protected EventClientRepository eventClientRepository;
-
-    @Before
-    public void setUp() throws JSONException {
-        mockStaticMethods();
-        rdtFormFragmentView = spy(new PatientRegisterFragmentStub());
-        presenter = new RDTJsonFormFragmentPresenter(rdtFormFragmentView, mock(RDTJsonFormInteractor.class));
-        Whitebox.setInternalState(presenter, "rdtJsonFormFragmentInteractor", interactor);
-    }
 
     @Test
     public void testPerformNextButtonActionShouldNavigateToNextStepAndSaveFormFromExpirationPage() {
@@ -160,7 +104,6 @@ public class RDTJsonFormFragmentPresenterTest {
         addMockStepDetails();
         mockStaticClasses();
         WeakReference<JsonFormFragmentView<JsonFormFragmentViewState>> viewRef = mock(WeakReference.class);
-        view = mock(JsonFormFragmentView.class);
         doReturn(RDT_TEST_JSON_FORM).when(view).getCurrentJsonState();
         doReturn(view).when(viewRef).get();
         Whitebox.setInternalState(presenter, "viewRef", viewRef);
@@ -171,42 +114,6 @@ public class RDTJsonFormFragmentPresenterTest {
         mStepDetails.put("next", "step1");
         Whitebox.setInternalState(presenter, "mStepDetails", mStepDetails);
     }
-
-    protected void mockStaticClasses() throws JSONException {
-        mockStatic(RDTJsonFormFragment.class);
-        formFragment = mock(RDTJsonFormFragment.class);
-        when(RDTJsonFormFragment.getFormFragment(anyString())).thenReturn(formFragment);
-    }
-
-    protected void mockStaticMethods() throws JSONException {
-        // mock RDTApplication and Drishti context
-        mockStatic(RDTApplication.class);
-        PowerMockito.when(RDTApplication.getInstance()).thenReturn(rdtApplication);
-        PowerMockito.when(rdtApplication.getStepStateConfiguration()).thenReturn(stepStateConfig);
-        PowerMockito.when(rdtApplication.getContext()).thenReturn(drishtiContext);
-
-        // mock repositories
-        PowerMockito.when(drishtiContext.getEventClientRepository()).thenReturn(eventClientRepository);
-
-        JSONObject jsonObject = mock(JSONObject.class);
-        doReturn("step1").when(jsonObject).optString(AdditionalMatchers.or(eq(SCAN_CARESTART_PAGE), eq(SCAN_QR_PAGE)));
-        doReturn("step1").when(jsonObject).optString(eq(PRODUCT_EXPIRED_PAGE), anyString());
-        doReturn("step6").when(jsonObject).optString(eq(PRODUCT_EXPIRED_PAGE));
-        doReturn("step9").when(jsonObject).optString(eq(BLOT_PAPER_TASK_PAGE));
-        doReturn("step1:expiration_date_reader").when(jsonObject).optString(eq(EXPIRATION_DATE_READER_ADDRESS), anyString());
-        doReturn("step1").when(jsonObject).optString(eq(TAKE_IMAGE_OF_RDT_PAGE));
-        doReturn("rdt_id").when(jsonObject).optString(eq(RDT_ID_KEY));
-        doReturn(new JSONArray("[\n" +
-                "    \"step7:lbl_rdt_id\",\n" +
-                "    \"step8:lbl_rdt_id\",\n" +
-                "    \"step9:lbl_rdt_id\",\n" +
-                "    \"step18:lbl_rdt_id\",\n" +
-                "    \"step19:lbl_rdt_id\"\n" +
-                "  ]")).when(jsonObject).optJSONArray(eq(RDT_ID_LBL_ADDRESSES));
-        doReturn("step20").when(jsonObject).optString(eq(MANUAL_EXPIRATION_DATE_ENTRY_PAGE));
-        doReturn(jsonObject).when(stepStateConfig).getStepStateObj();
-    }
-
 
     private String getExpirationDatePage(boolean isExpired) {
         Date date = isExpired ? getDateWithOffset(-1) : getDateWithOffset(1);
