+++ rdt-open-reader/src/test/java/io/ona/rdt/widget/RDTLabelFactoryTest.java
@@ -4,37 +4,21 @@ import android.content.Context;
 import android.view.Gravity;
 import android.view.View;
 
-import com.vijay.jsonwizard.domain.WidgetArgs;
 import com.vijay.jsonwizard.fragments.JsonFormFragment;
 import com.vijay.jsonwizard.interfaces.CommonListener;
 import com.vijay.jsonwizard.views.CustomTextView;
 
-import org.json.JSONException;
 import org.json.JSONObject;
-import org.junit.Before;
 import org.junit.Test;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-import org.powermock.api.mockito.PowerMockito;
-import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.reflect.Whitebox;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import androidx.constraintlayout.widget.ConstraintLayout;
-import io.ona.rdt.PowerMockTest;
-import io.ona.rdt.activity.RDTJsonFormActivity;
-import io.ona.rdt.application.RDTApplication;
 import io.ona.rdt.callback.OnLabelClickedListener;
-import io.ona.rdt.fragment.RDTJsonFormFragment;
-import io.ona.rdt.presenter.RDTJsonFormFragmentPresenter;
 import io.ona.rdt.util.Constants;
-import io.ona.rdt.util.StepStateConfig;
 
-import static com.vijay.jsonwizard.constants.JsonFormConstants.KEY;
-import static com.vijay.jsonwizard.constants.JsonFormConstants.NEXT;
-import static io.ona.rdt.util.Constants.Step.MANUAL_EXPIRATION_DATE_ENTRY_PAGE;
 import static io.ona.rdt.util.Constants.Step.SCAN_CARESTART_PAGE;
 import static io.ona.rdt.util.Constants.Step.SCAN_QR_PAGE;
 import static io.ona.rdt.widget.RDTLabelFactory.CENTER_LABEL;
@@ -44,36 +28,11 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
-import static org.powermock.api.mockito.PowerMockito.mockStatic;
 
 /**
  * Created by Vincent Karuri on 20/11/2019
  */
-@PrepareForTest({RDTApplication.class})
-public class RDTLabelFactoryTest extends PowerMockTest {
-
-    private RDTLabelFactory rdtLabelFactory;
-
-    protected JSONObject jsonObject;
-
-    @Mock
-    private RDTJsonFormFragment jsonFormFragment;
-
-    @Mock
-    private RDTApplication rdtApplication;
-
-    @Mock
-    private StepStateConfig stepStateConfig;
-
-    @Mock
-    protected RDTJsonFormFragmentPresenter presenter;
-
-    @Before
-    public void setUp() {
-        MockitoAnnotations.initMocks(this);
-        rdtLabelFactory = new RDTLabelFactory();
-        jsonObject = new JSONObject();
-    }
+public class RDTLabelFactoryTest extends BaseRDTLabelFactoryTest {
 
     @Test
     public void testEnhanceLabels() throws Exception {
@@ -114,29 +73,7 @@ public class RDTLabelFactoryTest extends PowerMockTest {
         rdtLabelFactory.getViewsFromJson("step", mock(Context.class), mock(JsonFormFragment.class), mock(JSONObject.class), mock(CommonListener.class), false);
     }
 
-    protected void mockStaticMethods() throws JSONException {
-        // mock RDTApplication and Drishti context
-        mockStatic(RDTApplication.class);
-        PowerMockito.when(RDTApplication.getInstance()).thenReturn(rdtApplication);
-        PowerMockito.when(rdtApplication.getStepStateConfiguration()).thenReturn(stepStateConfig);
-
-        jsonObject.put(SCAN_CARESTART_PAGE, SCAN_CARESTART_PAGE);
-        jsonObject.put(SCAN_QR_PAGE, SCAN_QR_PAGE);
-        jsonObject.put(MANUAL_EXPIRATION_DATE_ENTRY_PAGE, MANUAL_EXPIRATION_DATE_ENTRY_PAGE);
-        doReturn(jsonObject).when(stepStateConfig).getStepStateObj();
-    }
-
-    protected WidgetArgs getWidgetArgs(String widgetKey) throws JSONException {
-        WidgetArgs widgetArgs = new WidgetArgs();
-        jsonObject.put(KEY, widgetKey);
-        jsonObject.put(NEXT, "step1");
-        RDTJsonFormActivity rdtJsonFormActivity = mock(RDTJsonFormActivity.class);
-        doReturn(rdtJsonFormActivity).when(jsonFormFragment).getRdtActivity();
-
-        doReturn(presenter).when(jsonFormFragment).getPresenter();
-
-        widgetArgs.withFormFragment(jsonFormFragment)
-                .withJsonObject(jsonObject);
-        return widgetArgs;
+    protected RDTLabelFactory getRdtLabelFactory() {
+        return new RDTLabelFactory();
     }
 }
