+++ rdt-open-reader/src/testCovidIndonesia/java/io/ona/rdt/util/CovidFormLauncherTest.java
@@ -6,6 +6,7 @@ import org.json.JSONException;
 
 import io.ona.rdt.domain.Patient;
 
+import static io.ona.rdt.util.Constants.Form.RDT_TEST_FORM;
 import static io.ona.rdt.util.CovidConstants.Form.SAMPLE_COLLECTION_FORM;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.times;
@@ -15,7 +16,7 @@ import static org.powermock.api.mockito.PowerMockito.mock;
 /**
  * Created by Vincent Karuri on 15/07/2020
  */
-public class CovidFormLauncherTest extends FormLauncherTest {
+public class CovidFormLauncherTest extends BaseFormLauncherTest {
 
     @Override
     protected RDTJsonFormUtils getFormUtils() {
@@ -29,7 +30,7 @@ public class CovidFormLauncherTest extends FormLauncherTest {
 
     @Override
     protected void launchForms(Activity activity, Patient patient) throws JSONException {
-        super.launchForms(activity, patient);
+        formLauncher.launchForm(activity, RDT_TEST_FORM, patient);
         formLauncher.launchForm(activity, SAMPLE_COLLECTION_FORM, patient);
     }
